// apps/api/src/app/riot-api/utils/riot-api.constants.ts
export enum RIOT_SERVERS {
  BR1 = 'br1',
  EUN1 = 'eun1',
  EUW1 = 'euw1',
  JP1 = 'jp1',
  KR = 'kr',
  LA1 = 'la1',
  LA2 = 'la2',
  ME1 = 'me1',
  NA1 = 'na1',
  OC1 = 'oc1',
  RU = 'ru',
  SG2 = 'sg2',
  TR1 = 'tr1',
  TW2 = 'tw2',
  VN2 = 'vn2',
}

export enum RIOT_SERVERS_LABEL {
  BR1 = 'Brasil',
  EUN1 = 'Europa Nórdica e Oriental',
  EUW1 = 'Europa Ocidental',
  JP1 = 'Japão',
  KR = 'Coreia',
  LA1 = 'América Latina Norte',
  LA2 = 'América Latina Sul',
  ME1 = 'Oriente Médio',
  NA1 = 'América do Norte',
  OC1 = 'Oceania',
  RU = 'Rússia',
  SG2 = 'Singapura, Malásia e Outros (SEA)',
  TR1 = 'Turquia',
  TW2 = 'Taiwan',
  VN2 = 'Vietnã',
}